Update: Bill Manager, Person's Date/Time of Entry
Update - Bill Manager - updated the decimal data type in the BillPayment Table, Excel export feature.
Update - Person's Date/Time of Entry - Updated the validation form for both Merrylands and Umina's Incident Report, making mandatory only when an incident is related to an Assault. We hid the Person's Date/Time of Entry when values are empty.