Update Bill Manager
Red has requested to update the Trading Date setup.

Previously all transactions processed from 6:00am to 8:59am are set to a trading date of a day before the date it was processed (i.e. Transaction processed at 10/12/2021 07:50am, the Trading Date value will be set to 09/12/2021).
After the update, all transactions processed from 6:00am to 8:59am are set to a trading date of the current date (i.e. Transaction processed at 10/12/2021 07:50am, the Trading Date value will be set to 10/12/2021).

Take note that this is only applied for all transactions processed from 11/12/2021 onwards.